% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciBet.R
\name{Gambler_R}
\alias{Gambler_R}
\title{Select informative genes of the training set in a supervised manner.}
\usage{
Gambler_R(test_r, prob_r,ret_tab=FALSE)
}
\arguments{
\item{test_r}{Rows should be cells and columns should be genes.}

\item{prob_r}{trained scibet model  Rows should be genes and columns should be cell types.The genes must be matched with test_r}

\item{ret_tab}{return list or matrix.}
}
\value{
Predicted cell types in the form of list or matrix.
}
\description{
Use trained scibet model to predict cell types of the test matrix.
}
