% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciBet.R
\name{Learn_R}
\alias{Learn_R}
\title{Train SciBet model and generate a training matrix for cell type prediction.}
\usage{
Learn_R(expr, geneset, k=1000)
}
\arguments{
\item{expr}{The reference expression dataframeThe expression dataframe, with rows being cells, and columns being genes. The last column should be "label".}

\item{geneset}{A user-defined set of genes for prediction.}

\item{k}{See [SelectGene()] for details.}

}
\value{
A training matrix.
}
\description{
Train SciBet model and generate a training matrix for cell type prediction.
}

